Fix socket code to use proper `ssize_t` instead of plain `int`
Might fix a bug or two, might also expose a few more bugs. Needs proper testing.


git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45200 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/monitor/monitor_binary.h
@@ -44,7 +44,7 @@ void monitor_binary_event_closed(void);
 
 void monitor_check_binary(void);
 
-int monitor_binary_receive(unsigned char *buffer, size_t buffer_length);
+ssize_t monitor_binary_receive(unsigned char *buffer, size_t buffer_length);
 int monitor_binary_transmit(const unsigned char *buffer, size_t buffer_length);
 int monitor_binary_get_command_line(void);
